--- a/app/models/chat_channel.rb
+++ b/app/models/chat_channel.rb
@@ -2,7 +2,7 @@ class ChatChannel < ApplicationRecord
   include AlgoliaSearch
   attr_accessor :current_user, :usernames_string
 
-  has_many :messages, dependent: :destroy
+  has_many :messages
   has_many :chat_channel_memberships, dependent: :destroy
   has_many :users, through: :chat_channel_memberships