--- a/app/models/comment.rb
+++ b/app/models/comment.rb
@@ -97,6 +97,19 @@ class Comment < ApplicationRecord
     end
   end
 
+  def self.users_with_number_of_comments(user_ids, before_date)
+    joins(:user).
+      select("users.username, COUNT(comments.user_id) AS number_of_comments").
+      where(user_id: user_ids).
+      where(arel_table[:created_at].gt(before_date)).
+      group(User.arel_table[:username]).
+      order("number_of_comments DESC")
+  end
+
+  def self.tree_for(commentable, limit = 0)
+    commentable.comments.includes(:user).arrange(order: "score DESC").to_a[0..limit - 1].to_h
+  end
+
   def self.trigger_index(record, remove)
     # record is removed from index synchronously in before_destroy_actions
     return if remove
@@ -108,26 +121,10 @@ class Comment < ApplicationRecord
     end
   end
 
-  def self.users_with_number_of_comments(user_ids, before_date)
-    joins(:user).
-      select("users.username, COUNT(comments.user_id) AS number_of_comments").
-      where(user_id: user_ids).
-      where(arel_table[:created_at].gt(before_date)).
-      group(User.arel_table[:username]).
-      order("number_of_comments DESC")
-  end
-
+  # this should remain public because it's called by AlgoliaSearch::AlgoliaJob in .trigger_index
   def remove_algolia_index
     remove_from_index!
     Search::RemoveFromIndexJob.perform_now("ordered_comments_#{Rails.env}", index_id)
-  end
-
-  def index_id
-    "comments-#{id}"
-  end
-
-  def self.tree_for(commentable, limit = 0)
-    commentable.comments.includes(:user).arrange(order: "score DESC").to_a[0..limit - 1].to_h
   end
 
   def path
@@ -151,6 +148,8 @@ class Comment < ApplicationRecord
   end
 
   def id_code_generated
+    # 26 is the conversion base
+    # eg. 1000.to_s(26) would be "1cc"
     id.to_s(26)
   end
 
@@ -164,7 +163,8 @@ class Comment < ApplicationRecord
     return "[deleted]" if deleted
 
     text = ActionController::Base.helpers.strip_tags(processed_html).strip
-    HTMLEntities.new.decode ActionController::Base.helpers.truncate(text, length: length).gsub("&#39;", "'").gsub("&amp;", "&")
+    truncated_text = ActionController::Base.helpers.truncate(text, length: length).gsub("&#39;", "'").gsub("&amp;", "&")
+    HTMLEntities.new.decode(truncated_text)
   end
 
   def video
@@ -179,14 +179,18 @@ class Comment < ApplicationRecord
     end
   end
 
-  def remove_notifications
-    Notification.remove_all_without_delay(notifiable_ids: id, notifiable_type: "Comment")
-  end
-
   private
+
+  def index_id
+    "comments-#{id}"
+  end
 
   def update_notifications
     Notification.update_notifications(self)
+  end
+
+  def remove_notifications
+    Notification.remove_all_without_delay(notifiable_ids: id, notifiable_type: "Comment")
   end
 
   def update_descendant_notifications