--- a/app/models/job_opportunity.rb
+++ b/app/models/job_opportunity.rb
@@ -1,14 +1,16 @@
 class JobOpportunity < ApplicationRecord
+  REMOTENESS_PHRASES = {
+    "on_premise" => "In Office",
+    "fully_remote" => "Fully Remote",
+    "remote_optional" => "Remote Optional",
+    "on_premise_flexible" => "Mostly in Office but Flexible"
+  }.freeze
+
   has_many :articles
-  validates :remoteness,
-            inclusion: { in: %w[on_premise fully_remote remote_optional on_premise_flexible] }
+
+  validates :remoteness, inclusion: { in: REMOTENESS_PHRASES.keys }
+
   def remoteness_in_words
-    phrases = {
-      "on_premise" => "In Office",
-      "fully_remote" => "Fully Remote",
-      "remote_optional" => "Remote Optional",
-      "on_premise_flexible" => "Mostly in Office but Flexible"
-    }
-    phrases[remoteness]
+    REMOTENESS_PHRASES[remoteness]
   end
 end