--- a/app/models/page.rb
+++ b/app/models/page.rb
@@ -36,7 +36,8 @@ class Page < ApplicationRecord
   end
 
   def unique_slug_including_users_and_orgs
-    errors.add(:slug, "is taken.") if User.find_by(username: slug) || Organization.find_by(slug: slug) || Podcast.find_by(slug: slug)
+    slug_exists = User.exists?(username: slug) || Organization.exists?(slug: slug) || Podcast.exists?(slug: slug)
+    errors.add(:slug, "is taken.") if slug_exists
   end
 
   def bust_cache