--- a/app/models/poll_skip.rb
+++ b/app/models/poll_skip.rb
@@ -7,6 +7,7 @@ class PollSkip < ApplicationRecord
   private
 
   def one_vote_per_poll_per_user
-    errors.add(:base, "cannot vote more than once in one poll") if poll.poll_votes.where(user_id: user_id).any? || poll.poll_skips.where(user_id: user_id).any?
+    already_voted = poll.poll_votes.where(user_id: user_id).any? || poll.poll_skips.where(user_id: user_id).any?
+    errors.add(:base, "cannot vote more than once in one poll") if already_voted
   end
 end