--- a/spec/factories/poll_skips.rb
+++ b/spec/factories/poll_skips.rb
@@ -1,10 +1,4 @@
 FactoryBot.define do
   factory :poll_skip do
-    # prompt_markdown { Faker::Hipster.words(number: 5) }
-    # factory :poll_with_options do
-    #   after(:create) do |poll|
-    #     create_list(:poll_option, poll: poll)
-    #   end
-    # end
   end
 end