--- a/spec/models/badge_achievement_spec.rb
+++ b/spec/models/badge_achievement_spec.rb
@@ -1,9 +1,9 @@
 require "rails_helper"
 
 RSpec.describe BadgeAchievement, type: :model do
-  describe "validations" do
-    subject { create(:badge_achievement) }
+  let_it_be(:achievement) { create(:badge_achievement) }
 
+  describe "validations" do
     it { is_expected.to belong_to(:user) }
     it { is_expected.to belong_to(:badge) }
     it { is_expected.to belong_to(:rewarder).class_name("User").optional }
@@ -11,12 +11,10 @@ RSpec.describe BadgeAchievement, type: :model do
   end
 
   it "turns rewarding_context_message_markdown into rewarding_context_message HTML" do
-    achievement = create(:badge_achievement)
     expect(achievement.rewarding_context_message).to include("</a>")
   end
 
   it "awards credits after create" do
-    achievement = create(:badge_achievement)
     expect(achievement.user.credits.size).to eq(5)
   end
 end