--- a/spec/models/badge_spec.rb
+++ b/spec/models/badge_spec.rb
@@ -1,9 +1,9 @@
 require "rails_helper"
 
 RSpec.describe Badge, type: :model do
-  describe "validations" do
-    subject { create(:badge) }
+  let_it_be(:badge) { create(:badge) }
 
+  describe "validations" do
     it { is_expected.to have_many(:users).through(:badge_achievements) }
     it { is_expected.to have_many(:badge_achievements) }
     it { is_expected.to validate_presence_of(:title) }