--- a/spec/models/block_spec.rb
+++ b/spec/models/block_spec.rb
@@ -1,8 +1,8 @@
 require "rails_helper"
 
 RSpec.describe Block, type: :model do
-  let(:user) { create(:user) }
-  let(:block) { described_class.new(user: user, input_html: "hello") }
+  let_it_be(:user) { create(:user) }
+  let_it_be(:block) { described_class.new(user: user, input_html: "hello") }
 
   it "creates processed_html after published!" do
     user.add_role(:super_admin)
@@ -11,8 +11,7 @@ RSpec.describe Block, type: :model do
   end
 
   it "is not valid without user" do
-    block.user = nil
-    expect(block).not_to be_valid
+    expect(described_class.new(user: nil)).not_to be_valid
   end
 
   it "is not valid with non-admin user" do