--- a/spec/models/collection_spec.rb
+++ b/spec/models/collection_spec.rb
@@ -1,12 +1,10 @@
 require "rails_helper"
 
 RSpec.describe Collection, type: :model do
-  let(:user) { create(:user) }
-  let(:collection) { create(:collection, :with_articles, user: user) }
+  let_it_be(:user) { create(:user) }
+  let_it_be(:collection) { create(:collection, :with_articles, user: user) }
 
   describe "validations" do
-    subject { described_class.new }
-
     it { is_expected.to belong_to(:user) }
     it { is_expected.to belong_to(:organization).optional }
     it { is_expected.to have_many(:articles) }
@@ -17,11 +15,10 @@ RSpec.describe Collection, type: :model do
   end
 
   describe ".find_series" do
-    let(:user) { create(:user) }
-    let(:series) { create(:collection, user: user) }
+    let_it_be(:other_user) { create(:user) }
+    let_it_be(:series) { collection }
 
     it "returns an existing series" do
-      series # the series has to be created before the following expect
       expect do
         expect(described_class.find_series(series.slug, series.user)).to eq(series)
       end.not_to change(described_class, :count)
@@ -29,13 +26,11 @@ RSpec.describe Collection, type: :model do
 
     it "creates a new series for a user if an existing one is not found" do
       slug = Faker::Books::CultureSeries.book
-      expect { described_class.find_series(slug, user) }.to change(described_class, :count).by(1)
+      expect { described_class.find_series(slug, other_user) }.to change(described_class, :count).by(1)
     end
 
     it "creates a new series with an existing slug for a new user" do
-      user = create(:user)
-      series # the series has to be created before the following expect
-      expect { described_class.find_series(series.slug, user) }.to change(described_class, :count).by(1)
+      expect { described_class.find_series(series.slug, other_user) }.to change(described_class, :count).by(1)
     end
   end
 
@@ -44,7 +39,7 @@ RSpec.describe Collection, type: :model do
       Timecop.freeze(DateTime.parse("2019/10/24")) do
         allow(collection.articles).to receive(:update_all)
         collection.touch_articles
-        expect(collection.articles).to have_received(:update_all).with(updated_at: Time.zone.now)
+        expect(collection.articles).to have_received(:update_all).with(updated_at: Time.current)
       end
     end
   end