--- a/spec/models/event_spec.rb
+++ b/spec/models/event_spec.rb
@@ -1,7 +1,7 @@
 require "rails_helper"
 
 RSpec.describe Event, type: :model do
-  let(:event) { create(:event) }
+  let(:event) { build(:event) }
 
   it "rejects title with over 90 characters" do
     event.title = Faker::Lorem.characters(number: 100)
@@ -14,16 +14,18 @@ RSpec.describe Event, type: :model do
   end
 
   it "rejects ends times that are earlier than start times" do
-    event.ends_at = 14.hours.ago
+    event.ends_at = event.starts_at - 1.minute
     expect(event).not_to be_valid
   end
 
   it "creates slug for published events" do
-    event.published = true
-    expect(event).to be_valid
+    event = build(:event, category: "ama", title: "yo", published: true)
+    event.validate!
+    expected_slug = "#{event.category}-#{event.title}-#{event.starts_at.strftime('%m-%d-%Y')}"
+    expect(event.slug).to eq(expected_slug)
   end
 
   it "triggers cache busting on save" do
-    expect { build(:event).save }.to have_enqueued_job.on_queue("events_bust_cache")
+    expect { event.save }.to have_enqueued_job.on_queue("events_bust_cache")
   end
 end