<!--- Prepend PR title with [WIP] if work in progress. Remove when ready for review. -->

<!--- For a timely review/response, please avoid force-pushing additional commits if your PR already received reviews or comments -->

## What type of PR is this? (check all applicable)

- [x] Refactor
- [ ] Feature
- [x] Bug Fix
- [ ] Documentation Update

## Description

The title is a bit boastful and in Travis it might not be reach the same difference but I think I've managed to reduce the speed of the model tests by relatively a lot, depends on how you look at it :D

```shell
➜  devto git:(master) bundle exec rspec spec/models/
Finished in 1 minute 49.03 seconds (files took 11.53 seconds to load)
739 examples

➜  devto git:(rhymes/faster-model-specs) bundle exec rspec spec/models/
Finished in 1 minute 2.56 seconds (files took 24.22 seconds to load)
731 examples
```

So if from the total time we subtract the loading time we have these numbers:

- master: 97.5 seconds
- this branch: 38.34 seconds

To achieve this I went down the following (long) route:

- read [TestProf: a good doctor for slow Ruby tests](https://evilmartians.com/chronicles/testprof-a-good-doctor-for-slow-ruby-tests)
- basically applied all the stuff they talked about, as some point I went down a rabbit hole of flame graphs and memory optimization but I quickly realized that most of the issues are about object creation, not memory occupation (though that's another avenue I might explore in the future, for the app, not for the tests)
- I had ups and downs along the road, as not all ideas were feasible
- I also quickly realized that I had to refactor the tests (after all they are code as well 🤷‍♂) if I wanted to improve the tests. An example of refactoring session:
   - open a model spec file
   - comment out everything in the but the first test
   - uncomment each test one by one and reorganize the specs and object creation
   - repeat, while keeping an eye on object creation ([Factory Doctor](https://test-prof.evilmartians.io/#/factory_doctor.md) here helped **A LOT**)

In the process I also learned a lot about our codebase and testing :D I also moved some public stuff in `private` in the models. I'm going to add a review to explain what I did in the model files.

I also removed some redundant tests, added some new ones along the way, moved some tests that were in user spec in the proper decorator.

## Related Tickets & Documents

This was my first endeavor in the optimization land: https://github.com/thepracticaldev/dev.to/pull/4882